requirements.txt
absl-py==0.9.0
airsim==1.2.8
-e git+https://gitlab.sitcore.net/nick.waytowich/multimodal_hitl.git@3a634aad3a88a1dacb630ae2b718a246ed1ffbd0#egg=airsim_env&subdirectory=airsim
alembic==1.4.2
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.11.0
appdirs==1.4.4
argcomplete==1.12.0
astunparse==1.6.3
atpublic==2.0
azure-core==1.7.0
azure-storage-blob==12.3.2
backcall==0.1.0
bcrypt==3.2.0
boto3==1.14.47
botocore==1.17.47
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
commonmark==0.9.1
configobj==5.0.6
crcmod==1.7
cryptography==3.0
cycler==0.10.0
databricks-cli==0.11.0
dataclasses==0.7
decorator==4.4.2
dictdiffer==0.8.1
distro==1.5.0
docker==4.2.2
docutils==0.15.2
dpath==2.0.1
dvc==1.6.0
entrypoints==0.3
Flask==1.1.2
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.7
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.17.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-storage==1.19.0
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
gorilla==0.3.0
grandalf==0.6
grpcio==1.29.0
gunicorn==20.0.4
gym==0.17.2
h5py==2.10.0
httplib2==0.18.1
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpath-ng==1.5.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
knack==0.7.2
lxml==4.5.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mlflow==1.10.0
mplcursors==0.3
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
msrest==0.6.17
nanotime==0.5.2
networkx==2.4
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
oss2==2.6.1
packaging==20.4
pandas==1.0.4
paramiko==2.7.1
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
ply==3.11
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.4
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pydot==1.4.1
PyDrive2==1.6.0
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
pygtrie==2.3.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-ignite==0.3.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
rich==5.2.1
rsa==4.4.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
shortuuid==1.0.1
shtab==1.3.1
six==1.15.0
sklearn==0.0
sklearn-pandas==1.8.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.6.3
tobii-research==1.7.0
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tornado==4.5.3
tqdm==4.46.1
traitlets==4.3.3
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
voluptuous==0.11.7
wcwidth==0.2.3
webdavclient3==3.14.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.1.0
